requirements.txt
fastapi
logtail-python
uvicorn
openai
pydantic
sqlalchemy
python-dotenv
psycopg2-binary
starlette
pyjwt
passlib
openai
httpx
celery
redis
jinja2
asyncpg
flower
python-multipart
aiosmtplib
pydantic[email]
fastapi-jwt-auth
pydantic-settings
fast-bitrix24
gspread
google-auth
pytz